moves ExpressionDataSources type to useExpressionDataSources file
