+++ src/__mocks__/getExpressionDataSourcesMock.ts
@@ -1,6 +1,8 @@
 import { getApplicationSettingsMock } from 'src/__mocks__/getApplicationSettingsMock';
 import { staticUseLanguageForTests } from 'src/features/language/useLanguage';
-import type { ExpressionDataSources } from 'src/features/expressions/utils';
+import type { IInstanceDataSources } from 'src/types/shared';
+import type { ExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
+import type { NodeFormDataSelector } from 'src/utils/layout/useNodeItem';
 
 export function getExpressionDataSourcesMock(): ExpressionDataSources {
   return {
@@ -12,14 +14,14 @@ export function getExpressionDataSourcesMock(): ExpressionDataSources {
     layoutSettings: { pages: { order: [] } },
     optionsSelector: () => ({ isFetching: false, options: [] }),
     applicationSettings: getApplicationSettingsMock(),
-    instanceDataSources: {} as any,
+    instanceDataSources: {} as IInstanceDataSources | null,
     authContext: null,
     devToolsIsOpen: false,
     devToolsHiddenComponents: 'hide',
     langToolsSelector: () => staticUseLanguageForTests(),
     currentLanguage: 'nb',
     isHiddenSelector: () => false,
-    nodeFormDataSelector: () => ({}) as any,
+    nodeFormDataSelector: (() => ({})) as unknown as NodeFormDataSelector,
     nodeDataSelector: () => {
       throw new Error('Not implemented: nodeDataSelector()');
     },
