+++ src/features/expressions/index.test.ts
@@ -2,7 +2,7 @@ import { NodeNotFoundWithoutContext } from 'src/features/expressions/errors';
 import { evalExpr } from 'src/features/expressions/index';
 import { ExprVal } from 'src/features/expressions/types';
 import type { ExprConfig } from 'src/features/expressions/types';
-import type { ExpressionDataSources } from 'src/features/expressions/utils';
+import type { ExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
 
 describe('Expressions', () => {
   it('should return default value if expression evaluates to null', () => {
