+++ src/features/expressions/utils.ts
@@ -2,41 +2,9 @@ import dot from 'dot-object';
 
 import { ExprRuntimeError, NodeNotFound, NodeNotFoundWithoutContext } from 'src/features/expressions/errors';
 import { prettyErrors } from 'src/features/expressions/prettyErrors';
-import type { AttachmentsSelector } from 'src/features/attachments/AttachmentsStorePlugin';
-import type { DevToolsHiddenComponents } from 'src/features/devtools/data/types';
 import type { ExprConfig, Expression } from 'src/features/expressions/types';
-import type { IUseLanguage } from 'src/features/language/useLanguage';
-import type { NodeOptionsSelector } from 'src/features/options/OptionsStorePlugin';
-import type { FormDataRowsSelector, FormDataSelector } from 'src/layout';
-import type { ILayoutSettings } from 'src/layout/common.generated';
-import type { IApplicationSettings, IAuthContext, IInstanceDataSources, IProcess } from 'src/types/shared';
 import type { BaseLayoutNode, LayoutNode } from 'src/utils/layout/LayoutNode';
 import type { LayoutPage } from 'src/utils/layout/LayoutPage';
-import type { Hidden, NodeDataSelector } from 'src/utils/layout/NodesContext';
-import type { DataModelTransposeSelector } from 'src/utils/layout/useDataModelBindingTranspose';
-import type { NodeFormDataSelector } from 'src/utils/layout/useNodeItem';
-import type { NodeTraversalSelectorLax } from 'src/utils/layout/useNodeTraversal';
-
-export interface ExpressionDataSources {
-  process?: IProcess;
-  instanceDataSources: IInstanceDataSources | null;
-  applicationSettings: IApplicationSettings | null;
-  formDataSelector: FormDataSelector;
-  formDataRowsSelector: FormDataRowsSelector;
-  attachmentsSelector: AttachmentsSelector;
-  layoutSettings: ILayoutSettings;
-  optionsSelector: NodeOptionsSelector;
-  authContext: Partial<IAuthContext> | null;
-  langToolsSelector: (node: LayoutNode | undefined) => IUseLanguage;
-  currentLanguage: string;
-  isHiddenSelector: ReturnType<typeof Hidden.useIsHiddenSelector>;
-  nodeFormDataSelector: NodeFormDataSelector;
-  nodeDataSelector: NodeDataSelector;
-  nodeTraversal: NodeTraversalSelectorLax;
-  transposeSelector: DataModelTransposeSelector;
-  devToolsIsOpen: boolean;
-  devToolsHiddenComponents: DevToolsHiddenComponents;
-}
 
 export interface PrettyErrorsOptions {
   config?: ExprConfig;
