+++ src/hooks/useSourceOptions.ts
@@ -5,10 +5,10 @@ import { getKeyWithoutIndexIndicators } from 'src/utils/databindings';
 import { transposeDataBinding } from 'src/utils/databindings/DataBinding';
 import { useExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
 import type { ExprVal, ExprValToActualOrExpr } from 'src/features/expressions/types';
-import type { ExpressionDataSources } from 'src/features/expressions/utils';
 import type { IOptionInternal } from 'src/features/options/castOptionsToStrings';
 import type { IOptionSource } from 'src/layout/common.generated';
 import type { LayoutNode } from 'src/utils/layout/LayoutNode';
+import type { ExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
 
 interface IUseSourceOptionsArgs {
   source: IOptionSource | undefined;
