+++ src/utils/layout/generator/NodeGenerator.tsx
@@ -24,7 +24,6 @@ import { Hidden, NodesInternal } from 'src/utils/layout/NodesContext';
 import { useExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
 import type { SimpleEval } from 'src/features/expressions';
 import type { ExprConfig, ExprResolved, ExprValToActual, ExprValToActualOrExpr } from 'src/features/expressions/types';
-import type { ExpressionDataSources } from 'src/features/expressions/utils';
 import type { CompDef } from 'src/layout';
 import type { FormComponentProps, SummarizableComponentProps } from 'src/layout/common.generated';
 import type {
@@ -41,6 +40,7 @@ import type { ChildClaim } from 'src/utils/layout/generator/GeneratorContext';
 import type { LayoutNode, LayoutNodeProps } from 'src/utils/layout/LayoutNode';
 import type { HiddenState } from 'src/utils/layout/NodesContext';
 import type { BaseRow, StateFactoryProps } from 'src/utils/layout/types';
+import type { ExpressionDataSources } from 'src/utils/layout/useExpressionDataSources';
 
 /**
  * A node generator will always be rendered when a component is present in a layout, even if the component
