+++ src/utils/layout/useExpressionDataSources.ts
@@ -15,7 +15,39 @@ import { Hidden, NodesInternal } from 'src/utils/layout/NodesContext';
 import { useDataModelBindingTranspose } from 'src/utils/layout/useDataModelBindingTranspose';
 import { useNodeFormDataSelector } from 'src/utils/layout/useNodeItem';
 import { useNodeTraversalSelectorLax } from 'src/utils/layout/useNodeTraversal';
-import type { ExpressionDataSources } from 'src/features/expressions/utils';
+import type { AttachmentsSelector } from 'src/features/attachments/AttachmentsStorePlugin';
+import type { DevToolsHiddenComponents } from 'src/features/devtools/data/types';
+import type { IUseLanguage } from 'src/features/language/useLanguage';
+import type { NodeOptionsSelector } from 'src/features/options/OptionsStorePlugin';
+import type { FormDataRowsSelector, FormDataSelector } from 'src/layout';
+import type { ILayoutSettings } from 'src/layout/common.generated';
+import type { IApplicationSettings, IAuthContext, IInstanceDataSources, IProcess } from 'src/types/shared';
+import type { LayoutNode } from 'src/utils/layout/LayoutNode';
+import type { NodeDataSelector } from 'src/utils/layout/NodesContext';
+import type { DataModelTransposeSelector } from 'src/utils/layout/useDataModelBindingTranspose';
+import type { NodeFormDataSelector } from 'src/utils/layout/useNodeItem';
+import type { NodeTraversalSelectorLax } from 'src/utils/layout/useNodeTraversal';
+
+export interface ExpressionDataSources {
+  process?: IProcess;
+  instanceDataSources: IInstanceDataSources | null;
+  applicationSettings: IApplicationSettings | null;
+  formDataSelector: FormDataSelector;
+  formDataRowsSelector: FormDataRowsSelector;
+  attachmentsSelector: AttachmentsSelector;
+  layoutSettings: ILayoutSettings;
+  optionsSelector: NodeOptionsSelector;
+  authContext: Partial<IAuthContext> | null;
+  langToolsSelector: (node: LayoutNode | undefined) => IUseLanguage;
+  currentLanguage: string;
+  isHiddenSelector: ReturnType<typeof Hidden.useIsHiddenSelector>;
+  nodeFormDataSelector: NodeFormDataSelector;
+  nodeDataSelector: NodeDataSelector;
+  nodeTraversal: NodeTraversalSelectorLax;
+  transposeSelector: DataModelTransposeSelector;
+  devToolsIsOpen: boolean;
+  devToolsHiddenComponents: DevToolsHiddenComponents;
+}
 
 export function useExpressionDataSources(): ExpressionDataSources {
   const instanceDataSources = useLaxInstanceDataSources();
